riskengine_web/requirements.txt
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==37.0.4
cytoolz==0.12.0
Deprecated==1.2.13
Django==4.1
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-redis==5.2.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
drf-yasg==1.21.3
future==0.18.2
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.2
mysqlclient==2.1.1
oss2==2.16.0
packaging==21.3
pycparser==2.21
pycryptodome==3.17
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.7.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
scrypt==0.8.20
six==1.16.0
sqlparse==0.4.3
supervisor==4.2.4
toolz==0.12.0
uritemplate==4.1.1
urllib3==1.26.15
uWSGI==2.0.21
wrapt==1.15.0
zope.event==4.6
zope.interface==6.0